+++ index.html
@@ -251,6 +251,9 @@
   </template>
   master
   <script src="/js/function.js"></script>
+  <script src="/js/util.js" type="module"></script>
+  <script src="/js/main.js" type="module"></script>
+  <script src="/js/names.js" type="module"></script>
 </body>
 
 </html>
