+++ js/data.js
@@ -1,0 +1,26 @@
+const NAMES = [
+  'БРАНОСТАЛЬФ',
+  'ФРЕДИД',
+  'ТУБРАНО',
+  'АКМО',
+  'ЭРГАР',
+];
+
+const MESSAGE = [
+  'Всё отлично!',
+  'В целом всё неплохо.Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра.В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают.Как можно было поймать такой неудачный момент ? !',
+];
+
+const DESCRIPTION = [
+  'Изображение красного ретро-автомобиля',
+  'Вид на горы',
+  'Заход солнца на берегу океана',
+  'Звездное небо с большой медведицей',
+  'Звездное небо с созведием стрельца',
+];
+
+export { NAMES, MESSAGE, DESCRIPTION };
